+++ client/tests/integration/discussions-test.js
@@ -19,7 +19,9 @@ module('Integration: Discussions', {
     TestHelper.setup(App);
 
     paper = make('paper_with_discussion', { phases: [], tasks: [] });
-    topic = make('topic_with_replies', { paperId: paper.id, title: 'Hipster Ipsum Dolor' });
+    topic = make(
+      'topic_with_replies',
+      { paperId: paper.id, title: 'Hipster Ipsum Dolor' });
 
     $.mockjax({url: '/api/admin/journals/authorization', status: 204});
     $.mockjax({url: /\/api\/papers\/\d+\/manuscript_manager/, status: 204});