+++ engines/tahi_standard_tasks/client/app/components/invitations-display.js
@@ -12,12 +12,11 @@ export default Ember.Component.extend({
 
   latestDecision: null,
   latestDecisionInvitations: computed(
-    'latestDecision', 'latestDecision.invitations',
-    'latestDecision.invitations.[]', function() {
-      const type = this.get('invitationType');
+    'latestDecision.invitations.@each.inviteeRole', function() {
+      const type = this.get('inviteeRole');
       if (this.get('latestDecision.invitations')) {
         return this.get('latestDecision.invitations')
-                   .filterBy('invitationType', type);
+                   .filterBy('inviteeRole', type);
       }
     }
   ),
@@ -27,14 +26,14 @@ export default Ember.Component.extend({
   }),
 
   previousDecisionsWithFilteredInvitations: computed(
-    'previousDecisions', 'previousDecisions.[]', function() {
+    'previousDecisions.@each.inviteeRole', function() {
       return this.get('previousDecisions').map(decision => {
         const allInvitations = decision.get('invitations');
-        const type = this.get('invitationType');
+        const type = this.get('inviteeRole');
 
         decision.set(
           'filteredInvitations',
-          allInvitations.filterBy('invitationType', type)
+          allInvitations.filterBy('inviteeRole', type)
         );
 
         return decision;